pyproject.toml
[project]
name = "portfolio-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0",
    "groq>=0.29.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
